alert component, soft delete department and location, delete safety grouping asset

+++ app/Http/Controllers/AssetClassController.php
@@ -95,6 +95,10 @@ class AssetClassController extends Controller
     {
         Gate::authorize('is-admin');
 
+        if ($asset_class->subClasses()->exists()) {
+            return back()->with('error', 'Gagal dihapus! Grouping Asset Class ini masih digunakan dalam Grouping Asset Sub Class.');
+        }
+
         $asset_class->delete();
 
         return redirect()->route('asset-class.index')->with('success', 'Data berhasil dihapus!');
